fix database type, dashboard, auto stop

+++ be/app/models/exercise_feedback.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey, Text
+from sqlalchemy import Column, Integer, String, ForeignKey, Text, Numeric
 from sqlalchemy.orm import relationship
 from app.core.database import Base
 
@@ -9,7 +9,7 @@ class ExerciseFeedback(Base):
     rep_id = Column(Integer, ForeignKey("exercise_reps.rep_id"))
     body_part = Column(String(100)) # ส่วนที่ทำผิด
     issue = Column(Text) # แจ้งเตือนอะไร
-    timestamp_in_video = Column(Integer) # เวลาที่ผิดในวิดีโอ วินาที
+    timestamp_in_video = Column(Numeric(5, 2)) # เวลาที่ผิดในวิดีโอ วินาที
     
     # Relationships
     rep = relationship("ExerciseRep", back_populates="feedbacks")
